Move extensions to other file
